Updated TableView "scrollable" unit test for [TIMOB-25667]
